refactor(chat): remove stale ai/gui/chat_dock_widget.py duplicate
`src/shared/python/ai/gui/chat_dock_widget.py` was a second, older fork of
the canonical chat dock widget. Repoint the two importers
(`shared/python/dashboard/launcher.py`, the mujoco_humanoid_golf
`__main__`) at the canonical `src.shared.python.chat` module and delete
the duplicate. Both call sites switched from `engine_context=` to
`app_context=` to match the canonical kwarg.

Part of #5170.
