+++ src/engines/physics_engines/mujoco/python/mujoco_humanoid_golf/__main__.py
@@ -395,9 +395,9 @@ def main() -> None:
 
     # Add AI Chat dock widget (connects to FastAPI chat server)
     try:
-        from src.shared.python.ai.gui.chat_dock_widget import ChatDockWidget
+        from src.shared.python.chat import ChatDockWidget
 
-        chat_dock = ChatDockWidget(engine_context="mujoco", parent=win)
+        chat_dock = ChatDockWidget(app_context="mujoco", parent=win)
         win.addDockWidget(QtCore.Qt.DockWidgetArea.RightDockWidgetArea, chat_dock)
     except ImportError:
         pass  # Server not running — engine works fine without chat
